Devel

+++ src/static_builder.py
@@ -92,7 +92,7 @@ def git_clone(repo_name, url, revision=None):
     repo_dir = f"{args.default_dir}/{repo_name}"
     os.system(f"rm -rf {repo_dir}")
     os.makedirs(repo_dir, exist_ok=True)
-    kvPrint("\nRepository Name",repo_name)
+    kvPrint("\n\nRepository Name",repo_name)
     # os.system(f"git clone --quiet -n {url} {repo_dir} ")
     run_execute(f"[git clone] {repo_name} ",f"git clone --quiet -n {url} {repo_dir} ")
     os.chdir(repo_dir)
@@ -134,7 +134,7 @@ def main():
         if os.path.isfile(version_info_file) is False:
             if os.path.isdir(args.default_dir) is False:
                 os.mkdir(args.default_dir)
-            run_execute("static_vesion_info.json not found.", f'cp /src/static_version_info.json {version_info_file}')
+            run_execute("static_version_info.json not found.", f'cp /src/static_version_info.json {version_info_file}')
 
         if os.path.isdir(args.output_dir) is False:
             os.mkdir(args.output_dir)
@@ -158,17 +158,23 @@ def main():
         cprint(f" {repo_name} ,  {url} , {revision} ")
         git_clone(repo_name, url, revision)
         if repo_name == "icon_rc":
-            run_execute(f"-- Build {repo_name}", f"cd {args.default_dir}/{repo_name} ;  curl -O https://dl.google.com/go/go1.12.7.linux-amd64.tar.gz &&" +
-                                                "rm -rf /usr/local/go && " +
-                                                "tar zxf go1.12.7.linux-amd64.tar.gz -C /usr/local/  && " +
-                                                "rm go1.12.7.linux-amd64.tar.gz && "+
+
+            if run_execute("Find go compiler", "which /usr/local/go/bin/go", status_check="No") != 0:
+                run_execute(f"-- Installing Go ", f"curl -O https://dl.google.com/go/go1.12.7.linux-amd64.tar.gz &&" +
+                            "export GOROOT=/usr/local/go &&" +
+                            "export PATH=$GOROOT/bin:$GOPATH:/src/:$PATH &&"+
+                            "rm -rf /usr/local/go && " +
+                            "tar zxf go1.12.7.linux-amd64.tar.gz -C /usr/local/  && " +
+                            "rm go1.12.7.linux-amd64.tar.gz")
+
+            run_execute(f"-- Build {repo_name}", f"cd {args.default_dir}/{repo_name} ; "+
                                                 "export GOPATH=/go  &&"+
                                                 "export GOROOT=/usr/local/go &&" +
                                                 "export PATH=$GOROOT/bin:$GOPATH:/src/:$PATH &&"+
                                                 f"git checkout {revision}  && "+
                                                 "make linux && "+
                                                 f"make install DST_DIR={args.output_dir} && "+
-                                                "cd .. && rm -rf rewardcalculator /usr/local/go ;")
+                                                "cd .. && rm -rf rewardcalculator ")
         else:
             work_path = f"{args.default_dir}/{repo_name}"
             if args.verbose > 1:
@@ -177,7 +183,7 @@ def main():
                 quiet_mode = "-q"
             run_execute(f"Build {repo_name}", f"work_path={work_path}")
             run_execute(f"Install python dependencies", f"pip3 install {quiet_mode} -r requirements.txt", cwd=work_path)
-            run_execute(f"Build a wheel file", f"export VERSION={revision}; python3 setup.py bdist_wheel {quiet_mode} --dist-dir {args.output_dir}", cwd=work_path)
+            run_execute(f"Build a wheel file", f"export VERSION={revision}; python3 setup.py bdist_wheel {quiet_mode} --dist-dir {args.output_dir} >>/dev/null", cwd=work_path)
